% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_collinearity}
\alias{check_collinearity}
\title{Check for Collinearity Warning}
\usage{
check_collinearity(pos_coef, neg_coef, var_names, tol = 1e-06)
}
\arguments{
\item{pos_coef}{Positive coefficients vector.}

\item{neg_coef}{Negative coefficients vector.}

\item{var_names}{Variable names.}

\item{tol}{Tolerance for considering a coefficient as non-zero.}
}
\value{
A list with warning information.
}
\description{
Checks if any variable has both positive and negative coefficients non-zero.
}
\keyword{internal}
