% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{.advise_on_settings}
\alias{.advise_on_settings}
\title{Settings advice printed at fit start}
\usage{
.advise_on_settings(bootstrap, validation, refit, is_mids, verbose = TRUE)
}
\description{
Settings advice printed at fit start
}
\keyword{internal}
