% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_functions.R
\name{.compute_mi_unified_var_order}
\alias{.compute_mi_unified_var_order}
\title{Compute Unified Variable Order for sglwqs_mids Plots}
\usage{
.compute_mi_unified_var_order(
  object,
  top_n = 10,
  select_by = c("combined", "weight", "mi_selection_freq"),
  order_by = select_by,
  freq_col = "mi_selection_freq",
  require_nonzero_weight = TRUE,
  respect_groups = TRUE
)
}
\description{
Compute Unified Variable Order for sglwqs_mids Plots
}
\keyword{internal}
