% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weights.R
\name{.compute_weight_var_order}
\alias{.compute_weight_var_order}
\title{Compute Variable Order for Weight Plots}
\usage{
.compute_weight_var_order(object, sort_by = "weight", top_n = 10)
}
\description{
Compute Variable Order for Weight Plots
}
\keyword{internal}
