% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{.emit_diagnostics}
\alias{.emit_diagnostics}
\title{Emit diagnostics as fit-time messages}
\usage{
.emit_diagnostics(diagnostics, verbose = TRUE)
}
\description{
Emit diagnostics as fit-time messages
}
\keyword{internal}
