% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_functions.R
\name{.summarize_pooled_bootstrap}
\alias{.summarize_pooled_bootstrap}
\title{Summarize Pooled Bootstrap Results for sglwqs_mids}
\usage{
.summarize_pooled_bootstrap(object, conf_level = 0.95)
}
\description{
Summarize Pooled Bootstrap Results for sglwqs_mids
}
\keyword{internal}
