% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{extract_pooled_weights}
\alias{extract_pooled_weights}
\title{Deprecated: Use extract_weights() instead}
\usage{
extract_pooled_weights(object, ...)
}
\description{
Deprecated: Use extract_weights() instead
}
\keyword{internal}
