% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{extract_weights_grouped}
\alias{extract_weights_grouped}
\title{Deprecated: Use extract_weights() instead}
\usage{
extract_weights_grouped(object, ...)
}
\description{
Deprecated: Use extract_weights() instead
}
\keyword{internal}
