% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop_methods.R
\name{plot.sglwqs_calibration}
\alias{plot.sglwqs_calibration}
\title{Plot Calibration Table}
\usage{
\method{plot}{sglwqs_calibration}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"sglwqs_calibration"}.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot Calibration Table
}
