% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_functions.R
\name{pool_sglwqs_results}
\alias{pool_sglwqs_results}
\title{Pool sglwqs Results Using Rubin's Rules}
\usage{
pool_sglwqs_results(fits, verbose = TRUE)
}
\arguments{
\item{fits}{List of sglwqs objects}

\item{verbose}{Logical}
}
\value{
A list with pooled results
}
\description{
Pool sglwqs Results Using Rubin's Rules
}
\keyword{internal}
