% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_functions.R
\name{pool_validation_rubin}
\alias{pool_validation_rubin}
\title{Pool Validation Results Using Rubin's Rules}
\usage{
pool_validation_rubin(fits, verbose = TRUE)
}
\description{
Pool Validation Results Using Rubin's Rules
}
\keyword{internal}
