% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop_methods.R
\name{print.sglwqs_validation_metrics}
\alias{print.sglwqs_validation_metrics}
\title{Print Validation Metrics}
\usage{
\method{print}{sglwqs_validation_metrics}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"sglwqs_validation_metrics"}.}

\item{...}{Unused.}
}
\description{
Print Validation Metrics
}
