% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_functions.R
\name{summary.sglwqs_mids}
\alias{summary.sglwqs_mids}
\title{Summary of sglwqs_mids Object}
\usage{
\method{summary}{sglwqs_mids}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"sglwqs_mids"}.}

\item{...}{Additional arguments passed to \code{summary()}.}
}
\description{
Summary of sglwqs_mids Object
}
