% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_generics.R
\name{summary_by_group}
\alias{summary_by_group}
\title{Deprecated: summary_by_group is no longer needed}
\usage{
summary_by_group(object, ...)
}
\description{
Use summary(fit) to see group summaries, or extract_weights(fit, by_group = TRUE)
to get a data frame with group information.
}
\keyword{internal}
