% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_functions.R
\name{summary_inference}
\alias{summary_inference}
\alias{summary_inference.sglwqs}
\alias{summary_inference.sglwqs_mids}
\title{Summary of Available Inference Results}
\usage{
summary_inference(object, conf_level = 0.95)

summary_inference.sglwqs(object, conf_level = 0.95)

summary_inference.sglwqs_mids(object, conf_level = 0.95)
}
\arguments{
\item{object}{A fitted \code{sglwqs} or \code{sglwqs_mids} object.}

\item{conf_level}{Confidence level for intervals (default: 0.95).}
}
\description{
Information-source agnostic summary of available inference results.
}
